update fragment package
