+++ app/src/main/java/com/insurance/chatapp/ui/chat/ChatFragment.kt
@@ -51,71 +51,4 @@ abstract class ChatFragment : Fragment() {
     }
 }
 
-class BottomUserFragment : ChatFragment() {
 
-    override val chatType: ChatType
-        get() = ChatType.BOTTOM
-
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-
-        adapter.submitList(listOf(
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-        ))
-    }
-}
-
-class TopUserFragment : ChatFragment() {
-
-    override val chatType: ChatType
-        get() = ChatType.TOP
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-
-        adapter.submitList(listOf(
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Sender(MessageUiModel("გამგზავნი", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-            ChatListItem.Receiver(MessageUiModel("მიმღები", 123)),
-        ))
-    }
-}
